requirements.txt
torch
numpy
pandas
Pillow